get_dummy_tracer -> DummyTracer()
