+++ tests/contrib/bottle/test_autopatch.py
@@ -3,7 +3,7 @@ import ddtrace
 import webtest
 
 from unittest import TestCase
-from tests.test_tracer import get_dummy_tracer
+from tests.test_tracer import DummyTracer
 from ...utils import assert_span_http_status_code
 
 from ddtrace import compat
@@ -18,7 +18,7 @@ class TraceBottleTest(TestCase):
     """
     def setUp(self):
         # provide a dummy tracer
-        self.tracer = get_dummy_tracer()
+        self.tracer = DummyTracer()
         self._original_tracer = ddtrace.tracer
         ddtrace.tracer = self.tracer
         # provide a Bottle app
