+++ tests/contrib/falcon/test_distributed_tracing.py
@@ -1,5 +1,5 @@
 from falcon import testing
-from tests.test_tracer import get_dummy_tracer
+from tests.test_tracer import DummyTracer
 
 from .app import get_app
 
@@ -12,7 +12,7 @@ class DistributedTracingTestCase(testing.TestCase):
     def setUp(self):
         super(DistributedTracingTestCase, self).setUp()
         self._service = 'falcon'
-        self.tracer = get_dummy_tracer()
+        self.tracer = DummyTracer()
         self.api = get_app(tracer=self.tracer)
 
     def test_distributred_tracing(self):
@@ -33,7 +33,7 @@ class DistributedTracingTestCase(testing.TestCase):
         assert traces[0][0].trace_id == 100
 
     def test_distributred_tracing_disabled(self):
-        self.tracer = get_dummy_tracer()
+        self.tracer = DummyTracer()
         self.api = get_app(tracer=self.tracer, distributed_tracing=False)
         headers = {
             'x-datadog-trace-id': '100',
