+++ tests/contrib/molten/test_molten_di.py
@@ -10,7 +10,7 @@ from molten import DependencyInjector
 from ddtrace import Pin
 from ddtrace.contrib.molten import patch, unpatch
 
-from tests.tracer.test_tracer import get_dummy_tracer
+from tests.dummy import DummyTracer
 
 
 class Settings(dict):
@@ -85,7 +85,7 @@ class TestMoltenDI(TestCase):
 
     def setUp(self):
         patch()
-        self.tracer = get_dummy_tracer()
+        self.tracer = DummyTracer()
         Pin.override(molten, tracer=self.tracer, service=self.TEST_SERVICE)
 
     def tearDown(self):
