+++ tests/contrib/mongoengine/test_backwards.py
@@ -3,7 +3,7 @@ ensure old interfaces exist and won't break things.
 """
 import mongoengine
 
-from tests.test_tracer import get_dummy_tracer
+from tests.test_tracer import DummyTracer
 from tests.contrib import config
 
 
@@ -15,7 +15,7 @@ class Singer(mongoengine.Document):
 def test_less_than_v04():
     # interface from < v0.4
     from ddtrace.contrib.mongoengine import trace_mongoengine
-    tracer = get_dummy_tracer()
+    tracer = DummyTracer()
 
     connect = trace_mongoengine(tracer, service='my-mongo-db', patch=False)
     connect(port=config.MONGO_CONFIG['port'])
