+++ tests/contrib/mysql/test_backwards_compatibility.py
@@ -1,11 +1,11 @@
 
 from ddtrace.contrib.mysql import get_traced_mysql_connection
-from tests.test_tracer import get_dummy_tracer
+from tests.test_tracer import DummyTracer
 from tests.contrib import config
 
 
 def test_pre_v4():
-    tracer = get_dummy_tracer()
+    tracer = DummyTracer()
     MySQL = get_traced_mysql_connection(tracer, service='my-mysql-server')
     conn = MySQL(**config.MYSQL_CONFIG)
     cursor = conn.cursor()
