+++ tests/contrib/pymemcache/test_client_mixin.py
@@ -9,7 +9,7 @@ from ddtrace.contrib.pymemcache.patch import patch, unpatch
 from ddtrace.ext import memcached as memcachedx, net
 from .utils import MockSocket
 
-from tests.test_tracer import get_dummy_tracer
+from tests.test_tracer import DummyTracer
 from ...base import override_config
 from ...utils import assert_is_measured
 
@@ -53,7 +53,7 @@ class PymemcacheClientTestCaseMixin(unittest.TestCase):
         unpatch()
 
     def make_client(self, mock_socket_values, **kwargs):
-        tracer = get_dummy_tracer()
+        tracer = DummyTracer()
         Pin.override(pymemcache, tracer=tracer)
         self.client = pymemcache.client.base.Client((TEST_HOST, TEST_PORT), **kwargs)
         self.client.sock = MockSocket(list(mock_socket_values))
