+++ tests/tracer/test_sampler.py
@@ -16,12 +16,12 @@ from ddtrace.sampler import RateSampler, AllSampler, RateByServiceSampler
 from ddtrace.span import Span
 
 from tests.utils import override_env
-from .test_tracer import get_dummy_tracer
+from .test_tracer import DummyTracer
 
 
 @pytest.fixture
 def dummy_tracer():
-    return get_dummy_tracer()
+    return DummyTracer()
 
 
 def assert_sampling_decision_tags(span, agent=None, limit=None, rule=None):
@@ -31,7 +31,7 @@ def assert_sampling_decision_tags(span, agent=None, limit=None, rule=None):
 
 
 def create_span(tracer=None, name='test.span', meta=None, *args, **kwargs):
-    tracer = tracer or get_dummy_tracer()
+    tracer = tracer or DummyTracer()
     if 'context' not in kwargs:
         kwargs['context'] = tracer.get_call_context()
     span = Span(tracer=tracer, name=name, *args, **kwargs)
@@ -60,7 +60,7 @@ class RateSamplerTest(unittest.TestCase):
 
     def test_sample_rate_deviation(self):
         for sample_rate in [0.1, 0.25, 0.5, 1]:
-            tracer = get_dummy_tracer()
+            tracer = DummyTracer()
             writer = tracer.writer
 
             tracer.sampler = RateSampler(sample_rate)
@@ -82,7 +82,7 @@ class RateSamplerTest(unittest.TestCase):
 
     def test_deterministic_behavior(self):
         """ Test that for a given trace ID, the result is always the same """
-        tracer = get_dummy_tracer()
+        tracer = DummyTracer()
         writer = tracer.writer
 
         tracer.sampler = RateSampler(0.5)
@@ -116,7 +116,7 @@ class RateByServiceSamplerTest(unittest.TestCase):
 
     def test_sample_rate_deviation(self):
         for sample_rate in [0.1, 0.25, 0.5, 1]:
-            tracer = get_dummy_tracer()
+            tracer = DummyTracer()
             writer = tracer.writer
             tracer.configure(sampler=AllSampler())
             # We need to set the writer because tracer.configure overrides it,
@@ -169,7 +169,7 @@ class RateByServiceSamplerTest(unittest.TestCase):
             },
         ]
 
-        tracer = get_dummy_tracer()
+        tracer = DummyTracer()
         tracer.configure(sampler=AllSampler())
         priority_sampler = tracer.priority_sampler
         for case in cases:
@@ -395,7 +395,7 @@ def test_sampling_rule_matches_exception():
 
 @pytest.mark.parametrize('sample_rate', [0.01, 0.1, 0.15, 0.25, 0.5, 0.75, 0.85, 0.9, 0.95, 0.991])
 def test_sampling_rule_sample(sample_rate):
-    tracer = get_dummy_tracer()
+    tracer = DummyTracer()
     rule = SamplingRule(sample_rate=sample_rate)
 
     iterations = int(1e4 / sample_rate)
@@ -412,7 +412,7 @@ def test_sampling_rule_sample(sample_rate):
 
 
 def test_sampling_rule_sample_rate_1():
-    tracer = get_dummy_tracer()
+    tracer = DummyTracer()
     rule = SamplingRule(sample_rate=1)
 
     iterations = int(1e4)
@@ -423,7 +423,7 @@ def test_sampling_rule_sample_rate_1():
 
 
 def test_sampling_rule_sample_rate_0():
-    tracer = get_dummy_tracer()
+    tracer = DummyTracer()
     rule = SamplingRule(sample_rate=0)
 
     iterations = int(1e4)
